wrgdv3: connect credentials to CredentialsContainer via stores
